backend/requirements.txt
fastapi
uvicorn
pyzbar
tldextract
joblib
numpy
pandas
scikit-learn
cryptography>=41.0.0
Pillow
jpegio
stepic
